Update colorizer.htm
In my understanding, the function colorValue will be invoked whenever the input is changed. So, if we input k-word color, the colorValue() is called k times until it gets the right color. If that is true, then it is a redundancy, and fortunately, we can easily fix it with this code's topic: React's refs: we only need to read the value inside of the html's input when the form is submitted (the final color!). I have fixed the code accordingly and test it.